move TRC/Chain container to internal subpackage
This should make the intention clear, that this will not be shared
outside of the trust store.

Also remove hashing from the interface, as it is premature optimization.
