+++ go/lib/infra/modules/trust/v2/BUILD.bazel
@@ -4,7 +4,6 @@ go_library(
     name = "go_default_library",
     srcs = [
         "db.go",
-        "decode.go",
         "inserter.go",
         "inspector.go",
         "provider.go",
@@ -20,8 +19,8 @@ go_library(
         "//go/lib/ctrl/cert_mgmt:go_default_library",
         "//go/lib/infra:go_default_library",
         "//go/lib/infra/modules/db:go_default_library",
+        "//go/lib/infra/modules/trust/v2/internal/decoded:go_default_library",
         "//go/lib/scrypto:go_default_library",
-        "//go/lib/scrypto/cert/v2:go_default_library",
         "//go/lib/scrypto/trc/v2:go_default_library",
         "//go/lib/serrors:go_default_library",
     ],
