+++ go/lib/infra/modules/trust/v2/internal/decoded/BUILD.bazel
@@ -1,0 +1,12 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    srcs = ["decode.go"],
+    importpath = "github.com/scionproto/scion/go/lib/infra/modules/trust/v2/internal/decoded",
+    visibility = ["//go/lib/infra/modules/trust/v2:__subpackages__"],
+    deps = [
+        "//go/lib/scrypto/cert/v2:go_default_library",
+        "//go/lib/scrypto/trc/v2:go_default_library",
+    ],
+)
